Update backend to the latest version

+++ src/main/java/com/dauphine/blogger/services/CategoryServiceImpl.java
@@ -2,26 +2,17 @@ package com.dauphine.blogger.services;
 
 import com.dauphine.blogger.models.Category;
 import com.dauphine.blogger.repositories.CategoryRepository;
-import com.dauphine.blogger.services.CategoryService;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
 @Service
 public class CategoryServiceImpl implements CategoryService {
-    private final List<Category> temporaryCategories;
     private final CategoryRepository repository;
-    public CategoryServiceImpl(CategoryRepository repository) {
-    	this.repository=repository;
-    
 
-    
-        temporaryCategories = new ArrayList<>();
-        temporaryCategories.add(new Category(UUID.randomUUID(), "my first category"));
-        temporaryCategories.add(new Category(UUID.randomUUID(), "my second category"));
-        temporaryCategories.add(new Category(UUID.randomUUID(), "my third category"));
+    public CategoryServiceImpl(CategoryRepository repository) {
+        this.repository = repository;
     }
 
     @Override
@@ -31,8 +22,7 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category getByID(UUID id) {
-        return repository.findById(id)
-        		.orElse(null);
+        return repository.findById(id).orElse(null);
     }
 
     @Override
@@ -40,24 +30,29 @@ public class CategoryServiceImpl implements CategoryService {
         Category category = new Category(UUID.randomUUID(), name);
         return repository.save(category);
     }
-
     @Override
     public Category updateName(UUID id, String newName) {
         Category category = getByID(id);
-        if (category == null) {
-        	return null;
+        if (category != null) {
+            category.setName(newName);
+            return repository.save(category);
         }
-        String name = null;
-		category.setName(name);
-        return repository.save(category);
+        return null;
     }
-     
 
     @Override
     public boolean deleteById(UUID id) {
-       repository.deleteById(id);
-       return true;
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
-	
+    @Override
+    public List<Category> getAllLikeName(String name) {
+        return repository.findALLLikeName(name);
+    }
 }
+
+
